+++ src/NamespaceMap.php
@@ -9,21 +9,17 @@ declare(strict_types=1);
 
 namespace DecodeLabs\Archetype;
 
-use ArrayIterator;
-use DecodeLabs\Glitch\Dumpable;
-use IteratorAggregate;
-
-/**
- * @implements IteratorAggregate<string>
- */
-class NamespaceMap implements
-    IteratorAggregate,
-    Dumpable
+class NamespaceMap
 {
     /**
-     * @var array<string, int>
+     * @var array<string, NamespaceList>
      */
     protected array $namespaces = [];
+
+    /**
+     * @var array<string, array<string>>
+     */
+    protected array $aliases = [];
 
     /**
      * Add namespace
@@ -31,10 +27,15 @@ class NamespaceMap implements
      * @return $this
      */
     public function add(
+        string $root,
         string $namespace,
         int $priority = 0
     ): static {
-        $this->namespaces[$namespace] = $priority;
+        if (!isset($this->namespaces[$root])) {
+            $this->namespaces[$root] = new NamespaceList();
+        }
+
+        $this->namespaces[$root]->add($namespace, $priority);
         return $this;
     }
 
@@ -42,9 +43,12 @@ class NamespaceMap implements
      * Has namespace?
      */
     public function has(
+        string $root,
         string $namespace
     ): bool {
-        return isset($this->namespaces[$namespace]);
+        return
+            isset($this->namespaces[$root]) &&
+            $this->namespaces[$root]->has($namespace);
     }
 
     /**
@@ -53,31 +57,85 @@ class NamespaceMap implements
      * @return $this
      */
     public function remove(
+        string $root,
         string $namespace
     ): static {
-        unset($this->namespaces[$namespace]);
+        if (isset($this->namespaces[$root])) {
+            $this->namespaces[$root]->remove($namespace);
+        }
+
         return $this;
     }
 
     /**
-     * Get iterator
-     *
-     * @return ArrayIterator<int, string>
+     * Add alias
      */
-    public function getIterator(): ArrayIterator
-    {
-        uasort($this->namespaces, function ($a, $b) {
-            return $a <=> $b;
-        });
+    public function addAlias(
+        string $interface,
+        string $alias
+    ): void {
+        if(!isset($this->aliases[$interface])) {
+            $this->aliases[$interface] = [];
+        }
 
-        return new ArrayIterator(array_keys($this->namespaces));
+        $this->aliases[$interface][$alias] = $alias;
     }
 
     /**
-     * Dump for glitch
+     * Has alias
      */
-    public function glitchDump(): iterable
-    {
-        yield 'values' => $this->namespaces;
+    public function hasAlias(
+        string $interface,
+        string $alias
+    ): bool {
+        return isset($this->aliases[$interface][$alias]);
+    }
+
+    /**
+     * Remove alias
+     */
+    public function removeAlias(
+        string $interface,
+        string $alias
+    ): void {
+        unset($this->aliases[$interface][$alias]);
+    }
+
+    /**
+     * Map namespace
+     */
+    public function map(
+        string $namespace
+    ): NamespaceList {
+        $output = new NamespaceList();
+        $this->applyMap($namespace, $output);
+
+        foreach($this->aliases[$namespace] ?? [] as $alias) {
+            $this->applyMap($alias, $output);
+        }
+
+        return $output;
+    }
+
+    protected function applyMap(
+        string $namespace,
+        NamespaceList $namespaces
+    ): NamespaceList {
+        $parts = explode('\\', $namespace);
+        $inner = [];
+        $namespaces->add($namespace, -1);
+
+        while(!empty($parts)) {
+            $root = implode('\\', $parts);
+
+            if (isset($this->namespaces[$root])) {
+                $mapTo = empty($inner) ? null : implode('\\', $inner);
+                $namespaces->import($this->namespaces[$root], $mapTo);
+            }
+
+            array_unshift($inner, array_pop($parts));
+        }
+
+        return $namespaces;
     }
 }